requirements.txt
Pyrogram
TgCrypto
pydantic
pydantic-settings
lingua-language-detector
